加入UIRoot

+++ Engine/Source/Performance/Profiler.cs
@@ -40,14 +40,14 @@ public static class Profiler
         bool active = true,
         uint color = 0,
         string text = null,
-        [CallerLineNumber] uint lineNumber = 0,
+        [CallerLineNumber] int lineNumber = 0,
         [CallerFilePath] string filePath = null,
         [CallerMemberName] string memberName = null)
     {
         using var filestr = GetCString(filePath, out var fileln);
         using var memberstr = GetCString(memberName, out var memberln);
         using var namestr = GetCString(zoneName, out var nameln);
-        var srcLocId = TracyAllocSrclocName(lineNumber, filestr, fileln, memberstr, memberln, namestr, nameln);
+        var srcLocId = TracyAllocSrclocName((uint)lineNumber, filestr, fileln, memberstr, memberln, namestr, nameln);
         var context = TracyEmitZoneBeginAlloc(srcLocId, active ? 1 : 0);
 
         if (text != null)
